refactor: split prompt audit helpers
